Change PDO::PARAM_NULL to resolve to SQLT_CHR
When inserting NULLs for columns on & off in the definition below the value entered was a string of "0"

```php
schema::create('on_off_messages', function (Blueprint $table) {
    $table->increments('id', 'device_status_bits_pk');
    $table->string('on', 64)->nullable();
    $table->string('off', 64)->nullable();
});
```

This change allows NULLs to be entered. It appears that NULLs are still entered into integer fields correctly too, though I would advise this is tested in your environment too!