Add use delcarations
By adding use declarations at the top Oci8 and Statement classes, we no longer
have to use fully-qualified class names in the rest of the code. This makes it
easier to change the class namespaces to fit a particular project since you'll
only need to change the namespace and use declarations in each of the files.